rename PasskeyFAQ component to LoginFAQ and implement FAQ content
